+++ l10n_th_withholding_tax_cert/readme/HISTORY.rst
@@ -2,3 +2,9 @@
 ~~~~~~~~~~~~~~~~~~~~~~~
 
 * Start of the history
+
+
+12.0.1.0.2 (2020-02-12)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* Additional to create withholding tax cert from payment, it now create from Journal Entry